requirements.txt
# Core ML Framework
torch==2.0.0
torchvision==0.15.1
torchaudio==2.0.1

# Transformers and Model Hub
transformers==4.34.1
huggingface-hub==0.17.3
timm==0.9.2

# Computer Vision
opencv-python==4.8.0.76
openslide-python==1.3.1
open-clip-torch==2.23.0

# Data Science
pandas==2.0.3
numpy==1.25.0
matplotlib==3.7.3
scikit-learn==1.3.0
scikit-image==0.22.0
scipy==1.14.1

# ML Utilities
einops==0.7.0
nystrom-attention==0.0.12
torchmetrics==1.1.0

# Experiment Tracking
wandb==0.15.8

# Data I/O
h5py==3.9.0